make more updates- reorganize colors.js
